Join Count Ratio #2333
Add Join Count Ratio in unique values map dialog:
1. add a check box for “join count ratio”
2. a drop down list with available spatial weights

+++ VarTools.h
@@ -137,7 +137,7 @@ public:
 	bool IsAnyTimeVariant();
 	bool IsAnySyncWithGlobalTm();
 	wxString GetStrForTmId(int time);
-	int GetVarsCount() { return vars.size(); }
+	int GetVarsCount() { return (int)vars.size(); }
 	
 	/** Get variable name string */
 	wxString GetName(int var);
